feat: Support existing bracket fill of custom bracket
