+++ foam-madness/View/Tournament/TournamentGamesView/CreateCustomView/CustomTypeView.swift
@@ -19,15 +19,16 @@ struct CustomTypeView: View {
     @State var tournament: Tournament
     @State var customType = CustomType.random
     @State private var numTeams: Int = 0
+    @State private var chosenBracketFile: String? // existing
     @Binding var showGames: Bool
     
     var body: some View {
-        VStack(spacing: 20) {
+        VStack(spacing: 10) {
             Text("How do you want to fill your custom bracket?")
             
             Picker("", selection: $customType) {
                 Text("Randomly").tag(CustomType.random)
-                if numTeams >= 64 {
+                if numTeams > 4 {
                     Text("Base on Existing").tag(CustomType.existing)
                 }
                 Text("Select All").tag(CustomType.selectAll)
@@ -37,19 +38,28 @@ struct CustomTypeView: View {
             .scaleEffect(1.4)
             
             descriptionLabel
+                .font(.footnote)
                 .foregroundColor(.secondary)
             
             if customType == CustomType.existing {
-                // TODO: Show and handle year and M/F selection?
-                Text("Would allow selecting bracket here")
+                SelectInitialBracketView(
+                    isSimulated: tournament.isSimulated,
+                    showCustomSelector: false,
+                    chosenBracketFile: $chosenBracketFile
+                )
+                .scaleEffect(0.9)
+                .aspectRatio(contentMode: .fit)
+                .background(commonBlue.opacity(0.2))
+                .cornerRadius(5.0)
             }
             
             Button(action: handleContinue, label: {
                 Text("Continue")
             })
             .buttonStyle(PrimaryButtonFullWidthStyle())
+            .scaleEffect(0.8)
         }
-        .padding([.leading, .trailing], 20)
+        .padding([.leading, .trailing], 10)
         .onAppear {
             numTeams = (tournament.games?.count ?? 0) + 1
         }
@@ -61,7 +71,7 @@ struct CustomTypeView: View {
         case .random:
             return Text("The bracket will be initially filled with \(numTeams) random teams" + standardEndText)
         case .existing:
-            return Text("You'll choose an existing bracket to fill the teams initially" + standardEndText)
+            return Text("You'll choose an existing bracket to fill the teams initially" + standardEndText + " Fewer than 64 teams uses original seeds, not real results of later rounds. Men/Women selection here only affects the bracket used, not probabilities.")
         case .selectAll:
             return Text("You'll need to choose all \(numTeams) teams for the bracket, and can edit seeds as well.")
         }
@@ -70,8 +80,10 @@ struct CustomTypeView: View {
     func handleContinue() {
         if customType == CustomType.random {
             BracketCreationController(context: viewContext).fillTournamentWithRandomTeams(tournament)
+        } else if customType == CustomType.existing {
+            BracketCreationController(context: viewContext).fillTournamentFromExistingCustom(tournament, chosenBracketFile ?? "")
         }
-        // TODO: Logic to use existing bracket, if selected
+        // selectAll has no pre-processing
         showGames = true
     }
 }
